requirements.txt
aiofiles==24.1.0
aiogram==3.17.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
aniemore==1.2.3
annotated-types==0.7.0
anyio==4.8.0
attrs==24.3.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
datasets==2.21.0
dill==0.3.8
distro==1.9.0
ffmpeg-python==0.2.0
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.6.1
future==1.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.27.1
idna==3.10
Jinja2==3.1.5
jiter==0.8.2
magic-filter==1.0.12
MarkupSafe==3.0.2
more-itertools==10.6.0
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.4.2
numpy==1.26.4
openai==1.59.7
openai-whisper==20230124
packaging==24.2
pandas==2.2.3
propcache==0.2.1
pyarrow==19.0.0
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
six==1.17.0
sniffio==1.3.1
soundfile==0.12.1
sympy==1.13.1
tokenizers==0.13.3
torch==2.5.1
torchaudio==2.5.1
tqdm==4.67.1
transformers==4.26.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
xxhash==3.5.0
yarl==1.18.3
